backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
requests
gevent
